+++ src/FactFoundry.Blazor.Charts/Components/BarChart.razor.cs
@@ -16,6 +16,7 @@ public partial class BarChart : ComponentBase
     [Parameter] public bool CrosshairTooltip { get; set; }
     [Parameter] public int Width { get; set; } = 600;
     [Parameter] public int Height { get; set; } = 300;
+    [Parameter] public bool Responsive { get; set; }
     [Parameter] public ChartTheme? Theme { get; set; }
     [CascadingParameter] private ChartTheme? CascadingTheme { get; set; }